Huge update from Zhu
This adds the files given to me by Zhu. There have been some modifications to
make the code compile. Most of it is to make it easier to read.

The nnnsim.cc file is empty and must be filled in as the serialization
process is still not clear to me.

The NNN Address serialization must also be checked as I basically copied it
from the NDN code.

There is a lot of confusion in the wire models and I will have to see if
making a few folders wouldn't make the whole repository a little cleaner.

For today this is enough.