Serialization complete!
This commit handles all the serialization of the the NULL, SO and DO packets.
I have noticed that a lot of functions repeat, so in a following commit, I will
probably have to separate the definitions in different files. Doing so will
also probably make it easier for future packets to be added.

Will think it over.