Clean up of Zhu's initial files
This is mostly a clean up of Zhu's files so that the process of reading
through the files is a little easier.

I think the current one level repository is too big, so I will try to break
it up into different folders next commit.

This version does not yet compile due to the lack of a NNST or NNPT.