Added the trie files
Had to reformat the NNN Entries and forwarding strategies to conform with the
Trie. I am really over my head.

With this experience, I think that the NNPT should be, in theory, easier to
create.

I need to add the timer to delete entries. That will require more changes than
I realize.