Names container
Realized that a 3N / NNN node should have a way to keep track of the names that is using, has been using. This is my first attempt at a class to create this functionality. So far the container itself is empty and the entry has only be briefly designed. Will require a lot more work.