Implementation of updateLeaseTime
Realized that I needed to update the lease time at some point for the NamesContainer. This is the implementation for this particular function.

I have wondered what would happen if I try to add something that is already in the NamesContainer but I have not yet found a particular way to test it.

The test file has been updated to show how to use the new functions.