NNST FaceMetric multi_index
Due to the changes in the FaceMetric multi_index, I have to make a lot of changes to the NNST code. I have thus far commented the sections that are giving me issues. However it is clear that I will have to change a lot of the NNST code because I have changed a primary structure to deal with out particular issues.

As always this continues to be interesting.