More NNST Entry tests
I think that with this, I am finally finished with the NNST Entry. Now for the fun part with the Trie.

This commit includes print by metric, updates and everything else. I have had to use replace in the multi_index_container and this will probably bite me later on.