NNST implementation
This are all the current methods I think are required for the NNST. I have yet to test out the implementation. I had to change the NNST Entry printing functions to be able to be used by the trie implementation.

Much to my surprise, the trie implementation used didn't require modification.