Fixed NNST Entry deletion
I had used the Add function in the NNST to insert the PoAs without taking into account that I could just as equally add only one PoA at a time. This lead to the particular bug where a single added PoA was never queued for deletion.

Searching for the bug I also made use of the NS_LOG_FUNCTION which should really be included in all of the heavy classes for debugging purposes.

With the changes, I was able to use the NNST Entry freely. However, since the NNST is the one making the scheduling, I will need to make a wrapper in the NNST for these methods.

The nnn-nnst-test.cc file has been modified to take into account the new methods I am using.