More NNST drama
I was forced to make the cleanExpire functions work on the NNST. This has meant a further rewrite of the NNST code and the NNST Entry code. I also checked the erase iterators in the Entry and found inconsistencies that have been updated.

The trie does not apparently permit me to have an empty trie, so I will have to reevaluate the code to see what I can do.

Updating leases is currently not supported as it requires a connection from the NNST which I haven't written yet.

The test code has been updated to reflect the changes.