TCP graphing script
Kovi found an error that made it impossible to graph the data rate of TCP/IP using the R scripts we had. I have renamed the prior script to reflect that it is for use only on ICN/NDN data and have added a file to be used for TCP/IP data.