Started making the NNN-NDN faces
I have decided to include two NamesContainers on a machine, one for keeping track of the node's names and one for keeping track of the leases.

The implementation for the NNN-NDN face has been started but is turning out a lot more difficult that I expected. As always this will be interesting.