NNN-NDN Face
I have gone forward and backwards with the NNN-NDN face and what is apparent is that I need to be able to use both functions. This will be very hard because of the way the code is handled. I have decided to try to implement a multiple inheritance approach. It does not look as promising as I hope.